+++ lib/rubocop/socketry.rb
@@ -6,6 +6,7 @@
 require_relative "socketry/version"
 require_relative "socketry/plugin"
 require_relative "socketry/layout/consistent_blank_line_indentation"
+require_relative "socketry/layout/block_delimiter_spacing"
 
 # @namespace
 module RuboCop