+++ releases.md
@@ -1,5 +1,9 @@
 # Releases
 
+## Unreleased
+
+  - Added `Layout/BlockDelimiterSpacing` cop to enforce consistent spacing before block delimiters.
+
 ## v0.1.0
 
-  - Initial implementaiton of `Layout/ConsistentBlankLineIndentation`.
+  - Initial implementation of `Layout/ConsistentBlankLineIndentation`.